fix: enhancement

+++ main.go
@@ -2,33 +2,38 @@ package main
 
 import (
 	"os"
+	"time"
 
 	"github.com/sirupsen/logrus"
 )
 
+// Url and path
 const URL = "http://200.152.38.155/CNPJ/"
 const PATH = "data"
 const PATH_TEMP = "data/temp"
-const MAX_GOROUTINES = 250
+
+// Download the partitions
+const MAX_GOROUTINES = 500
 const CHUNK_SIZE = 10_000_000 // 10mb
 
-func main() {
-	os.MkdirAll(PATH, 0755)
-	os.MkdirAll(PATH_TEMP, 0755)
+// Retry configuration
+const REQUEST_TIMEOUT_MINUTES = 30
+const MAX_RETRY = 5
 
-	logrus.SetLevel(logrus.InfoLevel)
-	formatter := &logrus.TextFormatter{
-		TimestampFormat:        "15:04:05",
-		FullTimestamp:          true,
-		DisableLevelTruncation: true,
-	}
-	logrus.SetFormatter(formatter)
+func main() {
+	tini := time.Now()
 
 	urls, err := ExtractUrls(URL)
 	if err != nil {
 		panic(err)
 	}
 
-	Downloads(urls)
-	os.RemoveAll(PATH_TEMP)
+	errs := Downloads(urls)
+	timer := time.Since(tini).Minutes()
+	if len(errs) == 0 {
+		os.RemoveAll(PATH_TEMP)
+		logrus.Infof("Downloaded %d files in %.2f minutes", len(urls), timer)
+	} else {
+		logrus.Warnf("Downloaded %d(%d total) files with errors in %.2f minutes", len(errs), len(urls), timer)
+	}
 }
